Clean up / ported new functions to daily-main
Prepped for pull into master
