+++ legacy/main.py
@@ -5,7 +5,7 @@
 # and daily actions (Sudoku and weather by default).
 # Written by Adafruit Industries.  MIT license.
 #
-# MUST BE RUN AS ROOT (due to GPIO access)
+# MUST BE RUN AS ROOTi (due to GPIO access)
 #
 # Required software includes Adafruit_Thermal, Python Imaging and PySerial
 # libraries. Other libraries used are part of stock Python install.
@@ -17,11 +17,8 @@
 from __future__ import print_function
 import RPi.GPIO as GPIO
 import subprocess, time, Image, socket
-import datetime
-import filewalker
 from Adafruit_Thermal import *
 
-print("Starting up....")
 ledPin       = 18
 buttonPin    = 23
 holdTime     = 2     # Duration for button hold (shutdown)
@@ -31,31 +28,16 @@ dailyFlag    = False # Set after daily trigger occurs
 lastId       = '1'   # State information passed to/from interval script
 printer      = Adafruit_Thermal("/dev/serial0", 19200, timeout=5)
 
-# set item path
-dir_path = os.path.dirname(os.path.realpath(__file__))
-item_folder = dir_path + '/items'
-
-# initialize item of the day
-picked_item = ""
-
-# Daytime to reset the printer to print the new message of the next day
-hrs  = 4
-mins = 0
 
 # Called when button is briefly tapped.  Invokes time/temperature script.
 def tap():
-  print("Button press detected.")
   GPIO.output(ledPin, GPIO.HIGH)  # LED on while working
-  #subprocess.call(["python", "collected_before.py"])
-  #print today's message again
-  if picked_item != "":
-      printer.printImage(png_from_item(picked_item), True)
+  subprocess.call(["python", "timetemp.py"])
   GPIO.output(ledPin, GPIO.LOW)
 
 
 # Called when button is held down.  Prints image, invokes shutdown process.
 def hold():
-  print("Button hold detected.")
   GPIO.output(ledPin, GPIO.HIGH)
   printer.printImage(Image.open('gfx/goodbye.png'), True)
   printer.feed(3)
@@ -67,7 +49,6 @@ def hold():
 # Called at periodic intervals (30 seconds by default).
 # Invokes twitter script.
 def interval():
-  print("Twitter polled.")
   GPIO.output(ledPin, GPIO.HIGH)
   p = subprocess.Popen(["python", "twitter.py", str(lastId)],
     stdout=subprocess.PIPE)
@@ -75,24 +56,14 @@ def interval():
   return p.communicate()[0] # Script pipes back lastId, returned to main
 
 
-# Called once per day (5:30am by default).
+# Called once per day (6:30am by default).
+# Invokes weather forecast and sudoku-gfx scripts.
 def daily():
-  print("First button press of the day...")
   GPIO.output(ledPin, GPIO.HIGH)
-  #subprocess.call(["python", "daily_message.py"])
-  printer.printImage(Image.open(png_from_item(picked_item)), True)
+  subprocess.call(["python", "forecast.py"])
+  subprocess.call(["python", "sudoku-gfx.py"])
   GPIO.output(ledPin, GPIO.LOW)
 
-# takes a file path and return ready-to-print png image, converting text files by invoking MessageClass
-def png_from_item(picked_item):
-    if picked_item.rsplit(".",1)[1] == "png":
-        print "Image printed..."
-        return Image.open(picked_item)
-    elif picked_item.rsplit(".",1)[1] == "txt":
-        textfile_handle = open(picked_item, "r")
-        textfile_output = "".join(textfile_handle.readlines())
-        random_font = ImageFont.truetype(random.choice(fonts), 14, encoding='unic')
-        return Message(textfile_output,random.choice(Message_Style.values())).BuildMessage(random_font)
 
 # Initialization
 
@@ -108,6 +79,7 @@ GPIO.output(ledPin, GPIO.HIGH)
 
 # Processor load is heavy at startup; wait a moment to avoid
 # stalling during greeting.
+time.sleep(30)
 
 # Show IP address (if network is available)
 try:
@@ -117,34 +89,23 @@ try:
 	printer.feed(3)
 except:
 	printer.boldOn()
-	printer.println('Network is unreachable. Waiting 30 seconds...')
+	printer.println('Network is unreachable.')
 	printer.boldOff()
-        time.sleep(30)
-        try:
-		s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-		s.connect(('8.8.8.8', 0))
-		printer.print('My IP address is ' + s.getsockname()[0])
-		printer.feed(3)
-	except:
-		printer.boldOn()
-		printer.println('Network problems still occuring. Exiting...')
-		printer.boldOff()
-		printer.feed(3)
-		exit(0)
+	printer.print('Connect display and keyboard\n'
+	  'for network troubleshooting.')
+	printer.feed(3)
+	exit(0)
 
-# Print greeting image.
-#printer.print('Started at ' + datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
-printer.printImage(Image.open('gfx/helloworld.png'), True)
+# Print greeting image
+printer.printImage(Image.open('gfx/hello.png'), True)
 printer.feed(3)
 GPIO.output(ledPin, GPIO.LOW)
-	
+
 # Poll initial button state and time
 prevButtonState = GPIO.input(buttonPin)
 prevTime        = time.time()
 tapEnable       = False
 holdEnable      = False
-
-print("Waiting for event....")
 
 # Main loop
 while(True):
@@ -168,40 +129,31 @@ while(True):
       # Yes.  Debounced press or release...
       if buttonState == True:       # Button released?
         if tapEnable == True:       # Ignore if prior hold()
-
-          # Once per day (currently set for 6:30am local time, or when script
-          # is first run, if after 5:30am), run forecast and sudoku scripts.
-          if dailyFlag == False:
-            daily()
-            dailyFlag = True
-          else:
-            tap()                     # Tap triggered (button released)
+          tap()                     # Tap triggered (button released)
           tapEnable  = False        # Disable tap and hold
           holdEnable = False
       else:                         # Button pressed
         tapEnable  = True           # Enable tap and hold actions
         holdEnable = True
 
-  l = time.localtime()
-  if (60 * l.tm_hour + l.tm_min) > (60 * hrs + mins) and (60 * l.tm_hour + l.tm_min) < (60 * hrs + mins + 2):
-    dailyFlag = False
-    #clear current item of the day by moving file to 'used' bin
-  ###if picked_item != "":
-  ###    move_to_bin(picked_item)    
-    # pick item of the day
-    picked_item = filewalker.pick_item(item_folder,date.today())
-    # print item of the day
-    printer.printImage(png_from_item(picked_item), True)
-    
-
   # LED blinks while idle, for a brief interval every 2 seconds.
   # Pin 18 is PWM-capable and a "sleep throb" would be nice, but
   # the PWM-related library is a hassle for average users to install
   # right now.  Might return to this later when it's more accessible.
-  if ((int(t) & 1) == 0) and ((t - int(t)) < 0.15) and dailyFlag == False:
+  if ((int(t) & 1) == 0) and ((t - int(t)) < 0.15):
     GPIO.output(ledPin, GPIO.HIGH)
   else:
     GPIO.output(ledPin, GPIO.LOW)
+
+  # Once per day (currently set for 6:30am local time, or when script
+  # is first run, if after 6:30am), run forecast and sudoku scripts.
+  l = time.localtime()
+  if (60 * l.tm_hour + l.tm_min) > (60 * 6 + 30):
+    if dailyFlag == False:
+      daily()
+      dailyFlag = True
+  else:
+    dailyFlag = False  # Reset daily trigger
 
   # Every 30 seconds, run Twitter scripts.  'lastId' is passed around
   # to preserve state between invocations.  Probably simpler to do an
