+++ legacy/sudoku-txt.py
